+++ packages/hooks/src/useAcceptOffer.ts
@@ -10,11 +10,7 @@ import useCheckOwnership from './useCheckOwnership'
 import { convertTx } from './utils/transaction'
 
 gql`
-  query OfferWithApprovalAndFill(
-    $offerId: UUID!
-    $taker: Address!
-    $quantity: Uint256!
-  ) {
+  query OfferWithApproval($offerId: UUID!, $taker: Address!) {
     offer(id: $offerId) {
       type
       makerAddress
@@ -25,7 +21,22 @@ gql`
         chainId
         collectionAddress
       }
-      fill(account: $taker, quantity: $quantity) {
+    }
+  }
+`
+
+gql`
+  mutation CreateOfferFillTransaction(
+    $offerId: String!
+    $accountAddress: Address!
+    $quantity: Uint256!
+  ) {
+    createOfferFillTransaction(
+      offerId: $offerId
+      accountAddress: $accountAddress
+      quantity: $quantity
+    ) {
+      transaction {
         ...Transaction
       }
     }
@@ -71,10 +82,9 @@ export default function useAcceptOffer(signer: Signer | undefined): [
 
       try {
         // fetch approval from api
-        const { offer } = await sdk.OfferWithApprovalAndFill({
+        const { offer } = await sdk.OfferWithApproval({
           offerId: id,
           taker: account.toLowerCase(),
-          quantity: quantity.toString(),
         })
         invariant(offer, ErrorMessages.OFFER_NOT_FOUND)
 
@@ -122,6 +132,14 @@ export default function useAcceptOffer(signer: Signer | undefined): [
             ? offer.takerAddress || account.toLowerCase()
             : offer.makerAddress
 
+        // fetch fill transaction
+        const { createOfferFillTransaction } =
+          await sdk.CreateOfferFillTransaction({
+            accountAddress: account.toLowerCase(),
+            offerId: id,
+            quantity: quantity.toString(),
+          })
+
         // fetch initial quantity
         const { quantity: initialQuantity } = await checkOwnership(
           assetId,
@@ -130,7 +148,9 @@ export default function useAcceptOffer(signer: Signer | undefined): [
 
         setActiveProcess(AcceptOfferStep.TRANSACTION_SIGNATURE)
         // sign and broadcast the transaction
-        const tx = await signer.sendTransaction(convertTx(offer.fill))
+        const tx = await signer.sendTransaction(
+          convertTx(createOfferFillTransaction.transaction),
+        )
         setTransactionHash(tx.hash)
         setActiveProcess(AcceptOfferStep.TRANSACTION_PENDING)
         console.info(`waiting for transaction with hash ${tx.hash}...`)